+++ lizard_task/views.py
@@ -4,7 +4,7 @@ API views not coupled to models.
 from django.views.decorators.csrf import csrf_exempt
 from django.http import HttpResponseRedirect
 
-from lizard_task.models import PeriodicTaskExt
+from lizard_task.models import SecuredPeriodicTask
 from lizard_map.views import AppView
 
 
@@ -17,7 +17,8 @@ class TasksView(AppView):
     msg = ""
 
     def tasks(self):
-        return PeriodicTaskExt.objects.filter(data_set__isnull=False)
+        return SecuredPeriodicTask.objects.all()
+        #return SecuredPeriodicTask.objects.filter(data_set__isnull=False)
 
     def get(self, request, *args, **kwargs):
         self.msg = request.GET.get('msg', '')
@@ -29,7 +30,7 @@ class TasksView(AppView):
         """
         pk = request.POST['task_pk']
         # Lizard security is at work here.
-        periodic_task_ext = PeriodicTaskExt.objects.get(pk=pk)
-        periodic_task_ext.send_task(username=request.user.username)
-        msg = "Taak '%s' is opgestart." % periodic_task_ext.task.name
+        periodic_task = SecuredPeriodicTask.objects.get(pk=pk)
+        periodic_task.send_task(username=request.user.username)
+        msg = "Taak '%s' is opgestart." % periodic_task.name
         return HttpResponseRedirect('./?msg=%s' % msg)